+++ libs/audio_generation.py
@@ -1,237 +1,15 @@
 import GPy
-from constants import kernels_list
+from .constants import kernels_list
 import numpy as np
 import ddsp
 from typing import Type, Tuple, List
 from librosa import hz_to_note, note_to_hz, stft
-from utils import mel_to_hz
-from constants import SILENCE
-from constants import boundaries_scale_exp_voiced, boundaries_global_harmonic_volume_initial_bias, boundaries_global_noise_volume_initial_bias, boundaries_global_f0_mel_variance, boundaries_global_f0_mel_initial_bias, boundaries_global_harmonic_envelope_variance, boundaries_global_harmonic_envelope_lengthscale, boundaries_global_harmonic_envelope_initial_bias, boundaries_harmonic_envelope_kernel_name, boundaries_global_noise_distribution_initial_bias, boundaries_global_noise_distribution_mode, boundaries_kernel_freq_name, boundaries_hn_cor, boundaries_local_volume_hn_variance, boundaries_local_volume_hn_K_name,boundaries_local_f0_mel_variance,boundaries_local_f0_mel_lengthscale,boundaries_local_f0_mel_kernel_name, boundaries_n_harmonics, boundaries_f0_quantize,boundaries_ir_diminin
-from synth_params import sample_global_params, sample_local_params, GlobalParams, LocalParams
-
-n_kernel_kinds = len(kernels_list)
-
-def kernel_sampler(rng=None, variance=None, lengthscale=None, period=None):
-    kernels = [
-        ["RBF", GPy.kern.RBF(1, variance, lengthscale)],  # variance, lengthscale
-        ["Exponential", GPy.kern.Exponential(1, variance, lengthscale)],  # variance, lengthscale
-        ["Cosine", GPy.kern.Cosine(1, variance, lengthscale)],  # variance, lengthscale
-        ["Linear", GPy.kern.Linear(1, variance)],  # variances
-        ["sde_Brownian", GPy.kern.sde_Brownian(1, variance)],  # variance
-        ["Poly", GPy.kern.Poly(1, variance, lengthscale, bias=1., order=3.)],  # variance, scale, bias, order
-        ["StdPeriodic", GPy.kern.StdPeriodic(1, variance, period, lengthscale)],  # variance, period, lengthscale
-        ["PeriodicExponential", GPy.kern.PeriodicExponential(1, variance, lengthscale, period)],  # variance, lengthscale, period
-    ]
-    k = rng.integers(0, n_kernel_kinds)
-    return kernels[k]
-
-def labelize(value, boundaries):
-    """
-        boundaries must be 1d list-like in an ascending order.
-    """
-    boundaries = sorted(boundaries)
-    multi_label = np.zeros(len(boundaries) + 1)
-    
-    for enum, b in enumerate(boundaries):
-        if value < b:
-            multi_label[enum] = 1.
-            return multi_label
-    multi_label[-1] = 1.
-    
-    return multi_label
-
-def labelize_kernel(k_name):
-    kernel_dict = {"RBF":0, "Exponential":1, "Cosine":2, "Linear":3, "sde_Brownian":4, "Poly":5, "StdPeriodic":7, "PeriodicExponential":8}
-    return kernel_dict[k_name]
-
-def sample_voiced_silence_durations(
-    rng: np.random.Generator,
-    repeat_times: int,
-    scale_exp_silent: float,
-    scale_exp_voiced: float,
-    scale_normal_silent: float,
-    scale_normal_voiced: float,
-    duration: float,
-    frames_per_sec: int,
-    n_frames: int
-) -> Tuple[List[int], List[int]]:
-    """
-    Sample the lengths of voiced and silence segments, measured in frames,
-    and return both as a tuple.
-
-    This function uses exponential distributions for the initial silent and
-    voiced segments, and then uses normal distributions for subsequent segments
-    while ensuring that durations remain positive. Finally, it appends an
-    additional silent segment specified by `duration`.
-
-    Args:
-        rng: A random number generator (NumPy's >=1.17 recommended interface).
-        repeat_times: The number of times to alternate between a silent
-            segment and a voiced segment.
-        scale_exp_silent: The scale parameter for the exponential distribution
-            used to sample the first silent segment (in seconds).
-        scale_exp_voiced: The scale parameter for the exponential distribution
-            used to sample the first voiced segment (in seconds).
-        scale_normal_silent: The standard deviation for the normal distribution
-            used to sample subsequent silent segments (in seconds).
-        scale_normal_voiced: The standard deviation for the normal distribution
-            used to sample subsequent voiced segments (in seconds).
-        duration: An additional duration in seconds to be appended as a final
-            silent segment after all repeats.
-        frames_per_sec: The number of frames per second (typically sample_rate // hop_size).
-        n_frames: The maximum total number of frames allowed.
-
-    Returns:
-        A tuple (voiced_durations, silent_durations), where each is a list of
-        segment lengths in frames. The final total of frames is capped at `n_frames`.
-    """
-    # verbose arrays
-    _silent_durations = []
-    _voiced_durations = []
-
-    # set by sec
-    for i in range(repeat_times):
-        if i == 0:
-            _silent_durations.append(rng.exponential(scale_exp_silent))
-            _voiced_durations.append(rng.exponential(scale_exp_voiced))
-            
-        else:
-            while True:
-                candidate_silent_duration = rng.normal(_silent_durations[-1], scale_normal_silent)
-                if candidate_silent_duration > 0:
-                    break
-            _silent_durations.append(candidate_silent_duration)
-            
-            while True:
-                candidate_voiced_duration = rng.normal(_voiced_durations[-1], scale_normal_voiced)
-                if candidate_voiced_duration > 0:
-                    break
-            _voiced_durations.append(candidate_voiced_duration)
-
-    # padding
-    _silent_durations.append(duration)
-
-    # change to frames
-    _silent_durations = [int(elm * frames_per_sec) for elm in _silent_durations]
-    _voiced_durations = [int(elm * frames_per_sec) for elm in _voiced_durations]
-
-    # arrays extracted from verbose arrays
-    silent_durations = []
-    voiced_durations = []
-    total_frames = 0
-
-    for i in range(repeat_times + 1):
-        silent_durations.append(_silent_durations[i])
-        total_frames += _silent_durations[i]
-        if total_frames >= n_frames:
-            silent_durations[-1] -= total_frames - n_frames
-            break
-            
-        voiced_durations.append(_voiced_durations[i])
-        total_frames += _voiced_durations[i]
-        if total_frames >= n_frames:
-            voiced_durations[-1] -= total_frames - n_frames
-            break
-    return voiced_durations, silent_durations
-
-def generate_continuous_targets(global_params: GlobalParams, local_params: LocalParams, harmonic_envelope_kernel_name: str, global_noise_distribution_mode:float, kernel_noise_freq_name:str, local_volume_hn_K_name:str, local_f0_mel_kernel_name:str)  -> np.ndarray:
-    """
-    Generate continuous targets based on the provided global and local parameters.
-
-    Args:
-        global_params: An instance of GlobalParams containing global parameters.
-        local_params: An instance of LocalParams containing local parameters.
-
-    Returns:
-        A numpy array representing the generated continuous targets.
-    """
-    targets = np.array([global_params.scale_exp_voiced, global_params.harmonic_volume_initial_bias, global_params.noise_volume_initial_bias, global_params.f0_mel_variance, global_params.f0_mel_initial_bias, global_params.harmonic_envelope_variance, global_params.harmonic_envelope_lengthscale, global_params.harmonic_envelope_initial_bias, 
-                        labelize_kernel(harmonic_envelope_kernel_name), global_params.noise_distribution_initial_bias, global_noise_distribution_mode, labelize_kernel(kernel_noise_freq_name), local_params.hn_cor, local_params.volume_hn_variance, labelize_kernel(local_volume_hn_K_name), 
-                        local_params.f0_mel_variance, local_params.f0_mel_lengthscale, labelize_kernel(local_f0_mel_kernel_name), global_params.n_harmonics, global_params.f0_quantize, global_params.ir_diminin])
-
-    return targets
-
-def generate_discrete_targets(global_params: GlobalParams, local_params: LocalParams, harmonic_envelope_kernel_name: str, global_noise_distribution_mode:float, kernel_freq_name:str, local_volume_hn_K_name:str, local_f0_mel_kernel_name:str) -> dict:
-    labels = {
-            # Duration of voiced segment
-            "label_scale_exp_voiced":
-                labelize(global_params.scale_exp_voiced, boundaries_scale_exp_voiced),
-            
-            # Sharpness of global harmonic volume
-            "label_global_harmonic_volume_initial_bias":
-                labelize(global_params.harmonic_volume_initial_bias, boundaries_global_harmonic_volume_initial_bias), 
-            
-            # Sharpness of global noise volume
-            "label_global_noise_volume_initial_bias":
-                labelize(global_params.noise_volume_initial_bias, boundaries_global_noise_volume_initial_bias),
-            
-            # Global F0 variation range
-            "label_global_f0_mel_variance":
-                labelize(global_params.f0_mel_variance, boundaries_global_f0_mel_variance),
-                    
-            # # Harmonic pitch sharpness
-            "label_global_f0_mel_initial_bias":
-                labelize(global_params.f0_mel_initial_bias, boundaries_global_f0_mel_initial_bias),
-            
-            # Harmonic envelope variation range
-            "label_global_harmonic_envelope_variance":
-                labelize(global_params.harmonic_envelope_variance, boundaries_global_harmonic_envelope_variance),
-            
-            # Harmonic envelope lengthscale
-            "label_global_harmonic_envelope_lengthscale":
-                labelize(global_params.harmonic_envelope_lengthscale, boundaries_global_harmonic_envelope_lengthscale),
-            
-            # Harmonic envelope sharpness
-            "label_global_harmonic_envelope_initial_bias":
-                labelize(global_params.harmonic_envelope_initial_bias, boundaries_global_harmonic_envelope_initial_bias),
-
-                    
-            # Harmonic envelopne kernel
-            "label_harmonic_envelope_kernel_name":
-                labelize(harmonic_envelope_kernel_name, boundaries_harmonic_envelope_kernel_name),
-                    
-            # Sharpness of the noise distribution along the frequency axis
-            "label_global_noise_distribution_initial_bias":
-                labelize(global_params.noise_distribution_initial_bias, boundaries_global_noise_distribution_initial_bias),
-                    
-            # Spectral centroid of the noise distribution
-            "label_global_noise_distribution_mode":
-                labelize(global_noise_distribution_mode, boundaries_global_noise_distribution_mode),
-                    
-            # Kernel type of the inharmonic envelope
-            "label_kernel_freq_name":
-                labelize(kernel_freq_name, boundaries_kernel_freq_name),
-                    
-            # Local volume correlation between harmonic and inharmonic components
-            "label_hn_cor":
-                labelize(local_params.hn_cor, boundaries_hn_cor),
-                    
-            # Local volume variance
-            "label_local_volume_hn_variance":
-                labelize(local_params.volume_hn_variance, boundaries_local_volume_hn_variance),
-                    
-            # Kernel type of the local volume   
-            "label_local_volume_hn_K_name":
-                labelize(local_volume_hn_K_name, boundaries_local_volume_hn_K_name),
-                    
-            # Local F0 variance
-            "label_local_f0_mel_variance":
-                labelize(local_params.f0_mel_variance, boundaries_local_f0_mel_variance),
-                    
-            "label_local_f0_mel_lengthscale":
-                labelize(local_params.f0_mel_lengthscale, boundaries_local_f0_mel_lengthscale),
-                    
-            "label_local_f0_mel_kernel_name":
-                labelize(local_f0_mel_kernel_name, boundaries_local_f0_mel_kernel_name),
-                            
-            "label_n_harmonics":
-                labelize(global_params.n_harmonics, boundaries_n_harmonics),
-            "label_f0_quantize":
-                labelize(global_params.f0_quantize, boundaries_f0_quantize),
-            "label_ir_diminin":
-                labelize(global_params.ir_diminin, boundaries_ir_diminin)
-        }
-    return labels
+from .utils import mel_to_hz
+from .constants import SILENCE
+from .synth_params import sample_global_params, sample_local_params
+from .kernels import kernel_sampler
+from .duration import sample_voiced_silence_durations
+from .labeling import generate_continuous_targets, generate_discrete_targets
 
 def generate_one_sample(rng: np.random.Generator, max_n_to_mix: int, n_samples:int, small_variance:float, duration:float, frames_per_sec:int, n_frames:int,  default_kernel: Type[GPy.kern.Kern], n_frequencies:int, noise_anchors:int, noise_vagueness:int, sample_rate:int, workid:int):
     n_to_mix = rng.integers(1, max_n_to_mix + 1) #the number of source audio to mix
